+++ engine/src/runtime/source/platform/dx11/dx11_utils.cpp
@@ -10,7 +10,8 @@ void dx11SetDebugObjectName(ID3D11DeviceChild*      resource,
                             const std::string_view& name) {
 #if (defined(_DEBUG)) && (GRAPHICS_DEBUGGER_OBJECT_NAME)
     HRESULT const ret = resource->SetPrivateData(WKPDID_D3DDebugObjectName,
-                                                 name.size(), name.data());
+                                                 static_cast<UINT>(name.size()),
+                                                 name.data());
 #else
     UNREFERENCED_PARAMETER(resource);
     UNREFERENCED_PARAMETER(name);
@@ -20,8 +21,9 @@ void dx11SetDebugObjectName(ID3D11DeviceChild*      resource,
 
 void dxgiSetDebugObjectName(IDXGIObject* object, const std::string_view& name) {
 #if (defined(_DEBUG)) && (GRAPHICS_DEBUGGER_OBJECT_NAME)
-    HRESULT const ret = object->SetPrivateData(WKPDID_D3DDebugObjectName,
-                                               name.size(), name.data());
+    HRESULT const ret =
+            object->SetPrivateData(WKPDID_D3DDebugObjectName,
+                                   static_cast<UINT>(name.size()), name.data());
 #else
     UNREFERENCED_PARAMETER(resource);
     UNREFERENCED_PARAMETER(name);
